refactor: changed directory structure to reduce files
