+++ src/index.js
@@ -1,16 +1,16 @@
 const router = (module.exports = require("express").Router());
 
-router.use("/bedbug-inspection", require("./bedbug-inspection/bbi.route"));
-router.use("/bedbug-treatment", require("./bedbug-treatment/bbt.route"));
-router.use("/bird-exclusion", require("./bird-exclusion/birdex.route"));
-router.use("/bird-trapping", require("./bird-trapping/bird.route"));
-router.use("/dewebbing", require("./dewebbing/dewebbing.route"));
-router.use("/exterior-rodent", require("./exterior-rodent/ext-rodent.route"));
-router.use("/honeycomb", require("./honeycomb/honeycomb.route"));
-router.use("/interior-rodent", require("./interior-rodent/int-rodent.route"));
-router.use("/mosquito", require("./mosquito/mosquito.route"));
-router.use("/roach-assessment", require("./roach-assessment/roach.route"));
-router.use("/rodent-exclusion", require("./rodent-exclusion/rodent.route"));
-router.use("/scorpion", require("./scorpion/scorpion.route"));
-router.use("/snake", require("./snake/snake.route"));
-router.use("/termite", require("./termite/termite.route"));
+router.use("/bedbug-inspection", require("./routes/bedbug-inspection.route"));
+router.use("/bedbug-treatment", require("./routes/bedbug-treatment.route"));
+router.use("/bird-exclusion", require("./routes/bird-exclusion.route"));
+router.use("/bird-trapping", require("./routes/bird-trapping.route"));
+router.use("/dewebbing", require("./routes/dewebbing.route"));
+router.use("/exterior-rodent", require("./routes/exterior-rodent.route"));
+router.use("/honeycomb", require("./routes/honeycomb.route"));
+router.use("/interior-rodent", require("./routes/interior-rodent.route"));
+router.use("/mosquito", require("./routes/mosquito.route"));
+router.use("/roach-assessment", require("./routes/roach.route"));
+router.use("/rodent-exclusion", require("./routes/rodent-exclusion.route"));
+router.use("/scorpion", require("./routes/scorpion.route"));
+router.use("/snake", require("./routes/snake.route"));
+router.use("/termite", require("./routes/termite.route"));
