+++ src/models/bedbug-treatment.model.js
