+++ src/models/bird-exclusion.model.js
