+++ src/models/bird-trapping.model.js
