+++ src/models/exterior-rodent.model.js
