+++ src/models/interior-rodent.model.js
