+++ src/models/scorpion.model.js
