+++ src/models/snake.model.js
