+++ src/routes/interior-rodent.route.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const Controller = require("../controller/controller");
-const Model = require("./roach.model");
+const Model = require("../models/interior-rodent.model");
 
 // Initializes router
 const router = express.Router();
